# backend-node/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tasks": {
      "get": {
        "operationId": "TasksController_list",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedToMe",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}": {
      "put": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{id}/complete": {
      "post": {
        "operationId": "TasksController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/reports/audit": {
      "get": {
        "operationId": "ReportsController_audit",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/reports/pipeline-summary": {
      "get": {
        "operationId": "ReportsController_pipelineSummary",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/api/reports/origination-trends": {
      "get": {
        "operationId": "ReportsController_originationTrends",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/clients/{clientId}/nfs": {
      "get": {
        "operationId": "NfsController_list",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "nfs"
        ]
      },
      "post": {
        "operationId": "NfsController_create",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNfsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "nfs"
        ]
      }
    },
...
  },
  "info": {
    "title": "PRDF LMS API",
    "description": "Loan origination and servicing API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          }
        },
        "required": [
          "applicationId",
          "title"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          }
        }
      },
      "CompleteTaskDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          }
        }
      },
...
      "CreateNfsDto": {
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string"
          },
          "supportType": {
            "type": "string"
          },
          "durationHours": {
            "type": "number"
          },
          "dateProvided": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "supportType",
          "durationHours",
          "dateProvided"
        ]
      }
    }
  }
}