# backend-node/src/tasks/dto/update-task.dto.ts
import { PartialType, OmitType } from '@nestjs/swagger';
import { CreateTaskDto } from './create-task.dto';

export class UpdateTaskDto extends PartialType(OmitType(CreateTaskDto, ['applicationId'] as const)) {}